Convert all source files to unix newlines in the repository
If windows newlines are desired, git can be configured to adapt
the newlines on checkout instead of storing them in windows
newlines style in the repository.

Previously the style seemed to be a mix with differing conventions
from file to file (188 files with unix newlines and 111 files
with windows newlines).