Disable frame skipping in the encoder by default
Frame skipping seems to be triggered quite often in actual video
content (e.g. skipping almost 10% of frames in some situations).
Even if the frame skipping is disabled, the rest of the rate control
seems to be able to constrain the output pretty well to the target
bitrate, so the frame skipping feels unnecessary.