Consistently use the predefined _WIN32 and _WIN64 defines
The _WIN32 and _WIN64 defines are always defined automatically
by the compiler (both MSVC and mingw), so there's no need to
specify them in the project files.

The WIN32 defines might still be defined even if it isn't
defined manually, since some of the windows platform headers
defines it as well.