Explicitly request 2 threads when testing slice threading
Previously the default value of iMultipleThreadIdc was 0,
which made sure threads were used for the test with multiple
slices (if run on a multicore machine). Now the default is 1,
so multiple threads has to be requested.

Explicitly request 2 threads, to make sure the threading code is
tested, even on machines with only one core.