Silence MSVC warnings about "unsafe" standard C functions in console apps
Within the libraries themselves, we use other wrapper functions to
make sure the functions are used safely, but this wrapper layer isn't
available outside of the library in the console apps. The usage of these
functions is safe here, so silence the warnings.

This gets rid of all warnings when building with MSVC in 32 bit mode,
making it much easier to spot new warnings.