Reorder sections in encConsole.vcproj
The MSVC GUI seems to want to write the file with data in the order
Debug|Win32, Debug|x64, Release|Win32, Release|x64. This causes a
large number of changes in the git diff even though the content
still basically is the same. Therefore do it in a separate commit,
to make sure that any later actual changes stay minimal.