Consistently use the right type for log prints
These fields were originally int64_t. Even though long long is
the same size as int64_t in all setups, some compilers treat
it as a different type than long long (in 64 bit environments,
int64_t is a long, not a long long, even if they are the same
size).

%"PRId64" is used for printing int64_t, while %lld is used
for printing long long.

This fixes build warnings with gcc.