Build using -Werror in travis unit tests
This makes all compiler warnings to be treated as errors.

This allows catching warnings that are introduced, that may only be
visible with some compilers.

This option is only added within travis unit tests, not during normal
builds, so there is no risk of unnecessarily breaking builds on
new/untested systems.